+++ imports/core/api/players/server/publications.js
@@ -4,7 +4,7 @@ import { savePlayerId } from "../../../startup/server/connections.js";
 Meteor.publish("playerInfo", function({ playerId }) {
   const playerExists = Players.find(playerId).count() > 0;
   if (playerExists) {
-    savePlayerId(this.connection.id, playerId);
+    savePlayerId(this.connection, playerId);
   }
   return Players.find(playerId);
 });