+++ imports/core/api/stages/methods.js
@@ -1,6 +1,14 @@
 import { ValidatedMethod } from "meteor/mdg:validated-method";
 import SimpleSchema from "simpl-schema";
+
 import { Stages } from "./stages.js";
+
+let callOnChange, playerIdForConn;
+if (Meteor.isServer) {
+  playerIdForConn = require("../../startup/server/connections.js")
+    .playerIdForConn;
+  callOnChange = require("../server/onchange").callOnChange;
+}
 
 export const updateStageData = new ValidatedMethod({
   name: "Stages.methods.updateData",
@@ -34,5 +42,17 @@ export const updateStageData = new ValidatedMethod({
     const modifier = append ? { $push: update } : { $set: update };
 
     Stages.update(stageId, modifier, { autoConvert: false });
+
+    if (Meteor.isServer) {
+      callOnChange({
+        playerId: playerIdForConn(this.connection),
+        stageId,
+        stage,
+        key,
+        value: val,
+        prevValue: stage.data && stage.data[key],
+        append
+      });
+    }
   }
 });