+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
 - Simple page titles to help distinguish tabs.
 - New `append()` method on players, playerRounds, playerStages, games, stages and rounds.
 - Added `set()` to games, rounds and stages, and `get()` to games.
+- Added `onSet()`, `onAppend()` and `onChange()` callbacks.
 
 ### Changed